Default cron_command to `puppet agent`
It would be bad to provide `cron_ensure => present` without `cron_command`
because the command would default to the title of the resource, `puppet` in
this case.

We could either validate the combination of params or, what I think is
easier, set a sensible default to `puppet agent`. That can then be
overridden with `puppet apply` should someone wish.